Add a soft affinity for non-GPU nodes
This is a solution proposed in https://issues.redhat.com/browse/RHODS-3074

This allows us to have pods always prefer non-GPU nodes unless a GPU is explicitly requested. As mentioned in the JIRA, this probably is what users expect. This shouldn't break anything, tested on my cluster and it worked. If all CPU nodes are full (or unable to schedule onto them, etc), it will go onto a GPU node. That can be explicitly disabled by the labeling mechanism